Reorganize linters and fixers
