+++ lint/fixer/statement-order.ts
@@ -4,8 +4,8 @@
 import fs from 'node:fs';
 
 import { BrowserName, CompatStatement } from '../../types/types.js';
-import { IS_WINDOWS } from '../../test/utils.js';
-import compareStatements from '../lib/compare-statements.js';
+import { IS_WINDOWS } from '../utils.js';
+import compareStatements from '../../scripts/lib/compare-statements.js';
 
 /**
  * Return a new "support_block" object whose support statements have
