+++ lint/linter/test-multiple-statements.ts
