+++ lint/linter/test-obsolete.test.ts
